Refactor - Create a pylint.message package
There is a lot of Message related class in Utils
this warrant the creation of a new package.

See also review for burst utils.py into a package here:
https://github.com/PyCQA/pylint/pull/2654#issuecomment-470748956
