+++ pylint/message/message_definition.py
@@ -6,7 +6,7 @@
 import sys
 
 from pylint.exceptions import InvalidMessageError
-from pylint.utils.constants import MSG_TYPES
+from pylint.message.constants import MSG_TYPES
 from pylint.utils.normalize_text import normalize_text
