+++ pylint/utils/constants.py
@@ -5,29 +5,9 @@
 
 import re
 
-MSG_STATE_CONFIDENCE = 2
-_MSG_ORDER = "EWRCIF"
-MSG_STATE_SCOPE_CONFIG = 0
-MSG_STATE_SCOPE_MODULE = 1
-
-# The line/node distinction does not apply to fatal errors and reports.
-_SCOPE_EXEMPT = "FR"
-
 # Allow stopping after the first semicolon/hash encountered,
 # so that an option can be continued with the reasons
 # why it is active or disabled.
 OPTION_RGX = re.compile(r"\s*#.*\bpylint:\s*([^;#]+)[;#]{0,1}")
 
 PY_EXTS = (".py", ".pyc", ".pyo", ".pyw", ".so", ".dll")
-
-MSG_TYPES = {
-    "I": "info",
-    "C": "convention",
-    "R": "refactor",
-    "W": "warning",
-    "E": "error",
-    "F": "fatal",
-}
-MSG_TYPES_LONG = {v: k for k, v in MSG_TYPES.items()}
-
-MSG_TYPES_STATUS = {"I": 0, "C": 16, "R": 8, "W": 4, "E": 2, "F": 1}
